+++ oap-server/oal-rt/src/main/java/org/apache/skywalking/oal/rt/parser/SourceColumn.java
@@ -35,9 +35,10 @@ public class SourceColumn {
     private String fieldSetter;
     private String fieldGetter;
     private final boolean groupByCondInTopN;
+    private final boolean attribute;
 
     public SourceColumn(String fieldName, String columnName, Class<?> type, boolean isID, int length,
-                        boolean groupByCondInTopN) {
+                        boolean groupByCondInTopN, boolean attribute) {
         this.fieldName = fieldName;
         this.columnName = columnName;
         this.type = type;
@@ -48,6 +49,7 @@ public class SourceColumn {
         this.fieldGetter = ClassMethodUtil.toGetMethod(fieldName);
         this.fieldSetter = ClassMethodUtil.toSetMethod(fieldName);
         this.groupByCondInTopN = groupByCondInTopN;
+        this.attribute = attribute;
     }
 
     public void setFieldName(String fieldName) {