+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/meter/function/latest/LatestFunction.java
@@ -22,6 +22,7 @@ import java.util.Objects;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import org.apache.skywalking.oap.server.core.Const;
 import org.apache.skywalking.oap.server.core.UnexpectedException;
 import org.apache.skywalking.oap.server.core.analysis.manual.instance.InstanceTraffic;
 import org.apache.skywalking.oap.server.core.analysis.meter.Meter;
@@ -37,6 +38,7 @@ import org.apache.skywalking.oap.server.core.storage.annotation.Column;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Entity;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Storage;
 import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
+import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 @MeterFunction(functionName = "latest")
 @ToString
@@ -65,6 +67,7 @@ public abstract class LatestFunction extends Meter implements AcceptableValue<Lo
 
     @Override
     public void accept(MeterEntity entity, Long value) {
+        decorate(entity);
         this.entityId = entity.id();
         this.serviceId = entity.serviceId();
         this.value = value;
@@ -89,6 +92,12 @@ public abstract class LatestFunction extends Meter implements AcceptableValue<Lo
         metrics.setTimeBucket(toTimeBucketInHour());
         metrics.setServiceId(getServiceId());
         metrics.setValue(getValue());
+
+        metrics.setAttr0(getAttr0());
+        metrics.setAttr1(getAttr1());
+        metrics.setAttr2(getAttr2());
+        metrics.setAttr3(getAttr3());
+        metrics.setAttr4(getAttr4());
         return metrics;
     }
 
@@ -99,6 +108,12 @@ public abstract class LatestFunction extends Meter implements AcceptableValue<Lo
         metrics.setTimeBucket(toTimeBucketInDay());
         metrics.setServiceId(getServiceId());
         metrics.setValue(getValue());
+
+        metrics.setAttr0(getAttr0());
+        metrics.setAttr1(getAttr1());
+        metrics.setAttr2(getAttr2());
+        metrics.setAttr3(getAttr3());
+        metrics.setAttr4(getAttr4());
         return metrics;
     }
 
@@ -114,6 +129,26 @@ public abstract class LatestFunction extends Meter implements AcceptableValue<Lo
 
         this.entityId = remoteData.getDataStrings(0);
         this.serviceId = remoteData.getDataStrings(1);
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(2))) {
+            setAttr0(remoteData.getDataStrings(2));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(3))) {
+            setAttr1(remoteData.getDataStrings(3));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(4))) {
+            setAttr2(remoteData.getDataStrings(4));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(5))) {
+            setAttr3(remoteData.getDataStrings(5));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(6))) {
+            setAttr4(remoteData.getDataStrings(6));
+        }
     }
 
     @Override
@@ -125,6 +160,11 @@ public abstract class LatestFunction extends Meter implements AcceptableValue<Lo
         remoteBuilder.addDataStrings(entityId);
         remoteBuilder.addDataStrings(serviceId);
 
+        remoteBuilder.addDataStrings(getAttr0() == null ? Const.EMPTY_STRING : getAttr0());
+        remoteBuilder.addDataStrings(getAttr1() == null ? Const.EMPTY_STRING : getAttr1());
+        remoteBuilder.addDataStrings(getAttr2() == null ? Const.EMPTY_STRING : getAttr2());
+        remoteBuilder.addDataStrings(getAttr3() == null ? Const.EMPTY_STRING : getAttr3());
+        remoteBuilder.addDataStrings(getAttr4() == null ? Const.EMPTY_STRING : getAttr4());
         return remoteBuilder;
     }
 
@@ -153,6 +193,12 @@ public abstract class LatestFunction extends Meter implements AcceptableValue<Lo
             metrics.setTimeBucket(((Number) converter.get(TIME_BUCKET)).longValue());
             metrics.setServiceId((String) converter.get(InstanceTraffic.SERVICE_ID));
             metrics.setEntityId((String) converter.get(ENTITY_ID));
+
+            metrics.setAttr0((String) converter.get(ATTR0));
+            metrics.setAttr1((String) converter.get(ATTR1));
+            metrics.setAttr2((String) converter.get(ATTR2));
+            metrics.setAttr3((String) converter.get(ATTR3));
+            metrics.setAttr4((String) converter.get(ATTR4));
             return metrics;
         }
 
@@ -162,6 +208,12 @@ public abstract class LatestFunction extends Meter implements AcceptableValue<Lo
             converter.accept(TIME_BUCKET, storageData.getTimeBucket());
             converter.accept(InstanceTraffic.SERVICE_ID, storageData.getServiceId());
             converter.accept(ENTITY_ID, storageData.getEntityId());
+
+            converter.accept(ATTR0, storageData.getAttr0());
+            converter.accept(ATTR1, storageData.getAttr1());
+            converter.accept(ATTR2, storageData.getAttr2());
+            converter.accept(ATTR3, storageData.getAttr3());
+            converter.accept(ATTR4, storageData.getAttr4());
         }
     }