+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/meter/function/max/MaxFunction.java
@@ -21,6 +21,7 @@ package org.apache.skywalking.oap.server.core.analysis.meter.function.max;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import org.apache.skywalking.oap.server.core.Const;
 import org.apache.skywalking.oap.server.core.UnexpectedException;
 import org.apache.skywalking.oap.server.core.analysis.manual.instance.InstanceTraffic;
 import org.apache.skywalking.oap.server.core.analysis.meter.Meter;
@@ -38,6 +39,7 @@ import org.apache.skywalking.oap.server.core.storage.type.Convert2Storage;
 import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
 
 import java.util.Objects;
+import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 @MeterFunction(functionName = "max")
 @ToString
@@ -67,6 +69,7 @@ public abstract class MaxFunction extends Meter implements AcceptableValue<Long>
 
     @Override
     public void accept(MeterEntity entity, Long value) {
+        decorate(entity);
         setEntityId(entity.id());
         setServiceId(entity.serviceId());
         if (this.value < value) {
@@ -95,6 +98,12 @@ public abstract class MaxFunction extends Meter implements AcceptableValue<Long>
         metrics.setTimeBucket(toTimeBucketInHour());
         metrics.setServiceId(getServiceId());
         metrics.setValue(getValue());
+
+        metrics.setAttr0(getAttr0());
+        metrics.setAttr1(getAttr1());
+        metrics.setAttr2(getAttr2());
+        metrics.setAttr3(getAttr3());
+        metrics.setAttr4(getAttr4());
         return metrics;
     }
 
@@ -105,6 +114,12 @@ public abstract class MaxFunction extends Meter implements AcceptableValue<Long>
         metrics.setTimeBucket(toTimeBucketInDay());
         metrics.setServiceId(getServiceId());
         metrics.setValue(getValue());
+
+        metrics.setAttr0(getAttr0());
+        metrics.setAttr1(getAttr1());
+        metrics.setAttr2(getAttr2());
+        metrics.setAttr3(getAttr3());
+        metrics.setAttr4(getAttr4());
         return metrics;
     }
 
@@ -119,6 +134,26 @@ public abstract class MaxFunction extends Meter implements AcceptableValue<Long>
         setServiceId(remoteData.getDataStrings(1));
         setTimeBucket(remoteData.getDataLongs(1));
         setValue(remoteData.getDataLongs(0));
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(2))) {
+            setAttr0(remoteData.getDataStrings(2));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(3))) {
+            setAttr1(remoteData.getDataStrings(3));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(4))) {
+            setAttr2(remoteData.getDataStrings(4));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(5))) {
+            setAttr3(remoteData.getDataStrings(5));
+        }
+
+        if (StringUtil.isNotEmpty(remoteData.getDataStrings(6))) {
+            setAttr4(remoteData.getDataStrings(6));
+        }
     }
 
     @Override
@@ -128,6 +163,12 @@ public abstract class MaxFunction extends Meter implements AcceptableValue<Long>
         remoteBuilder.addDataLongs(getTimeBucket());
         remoteBuilder.addDataStrings(getEntityId());
         remoteBuilder.addDataStrings(getServiceId());
+
+        remoteBuilder.addDataStrings(getAttr0() == null ? Const.EMPTY_STRING : getAttr0());
+        remoteBuilder.addDataStrings(getAttr1() == null ? Const.EMPTY_STRING : getAttr1());
+        remoteBuilder.addDataStrings(getAttr2() == null ? Const.EMPTY_STRING : getAttr2());
+        remoteBuilder.addDataStrings(getAttr3() == null ? Const.EMPTY_STRING : getAttr3());
+        remoteBuilder.addDataStrings(getAttr4() == null ? Const.EMPTY_STRING : getAttr4());
         return remoteBuilder;
     }
 
@@ -157,6 +198,12 @@ public abstract class MaxFunction extends Meter implements AcceptableValue<Long>
             metrics.setServiceId((String) converter.get(InstanceTraffic.SERVICE_ID));
             metrics.setTimeBucket(((Number) converter.get(TIME_BUCKET)).longValue());
             metrics.setValue(((Number) converter.get(VALUE)).longValue());
+
+            metrics.setAttr0((String) converter.get(ATTR0));
+            metrics.setAttr1((String) converter.get(ATTR1));
+            metrics.setAttr2((String) converter.get(ATTR2));
+            metrics.setAttr3((String) converter.get(ATTR3));
+            metrics.setAttr4((String) converter.get(ATTR4));
             return metrics;
         }
 
@@ -166,6 +213,12 @@ public abstract class MaxFunction extends Meter implements AcceptableValue<Long>
             converter.accept(InstanceTraffic.SERVICE_ID, storageData.getServiceId());
             converter.accept(TIME_BUCKET, storageData.getTimeBucket());
             converter.accept(VALUE, storageData.getValue());
+
+            converter.accept(ATTR0, storageData.getAttr0());
+            converter.accept(ATTR1, storageData.getAttr1());
+            converter.accept(ATTR2, storageData.getAttr2());
+            converter.accept(ATTR3, storageData.getAttr3());
+            converter.accept(ATTR4, storageData.getAttr4());
         }
     }