+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/SourceReceiverImpl.java
@@ -18,17 +18,24 @@
 
 package org.apache.skywalking.oap.server.core.source;
 
+import com.google.common.collect.ImmutableSet;
+import com.google.common.reflect.ClassPath;
 import java.io.IOException;
 import lombok.Getter;
 import org.apache.skywalking.oap.server.core.analysis.DispatcherDetectorListener;
 import org.apache.skywalking.oap.server.core.analysis.DispatcherManager;
+import org.apache.skywalking.oap.server.core.analysis.SourceDecoratorManager;
 
 public class SourceReceiverImpl implements SourceReceiver {
     @Getter
     private final DispatcherManager dispatcherManager;
 
+    @Getter
+    private final SourceDecoratorManager sourceDecoratorManager;
+
     public SourceReceiverImpl() {
         this.dispatcherManager = new DispatcherManager();
+        this.sourceDecoratorManager = new SourceDecoratorManager();
     }
 
     @Override
@@ -42,6 +49,12 @@ public class SourceReceiverImpl implements SourceReceiver {
     }
 
     public void scan() throws IOException, InstantiationException, IllegalAccessException {
-        dispatcherManager.scan();
+        ClassPath classpath = ClassPath.from(this.getClass().getClassLoader());
+        ImmutableSet<ClassPath.ClassInfo> classes = classpath.getTopLevelClassesRecursive("org.apache.skywalking");
+        for (ClassPath.ClassInfo classInfo : classes) {
+            Class<?> aClass = classInfo.load();
+            sourceDecoratorManager.addIfAsSourceDecorator(aClass);
+            dispatcherManager.addIfAsSourceDispatcher(aClass);
+        }
     }
 }